Move all QgsTextFormat tests to correct file
